+++ src/lib/visual-editor/actions/changeConfig.ts
@@ -1,7 +1,7 @@
 /* eslint-disable functional/no-return-void */
-import { SDKConfig } from '../msg-router/incoming-types';
 import * as mutators from '../mutators';
 import { buildLocalStyles } from '../mutators/localStyles';
+import { SDKConfig } from '../types';
 import { Mode } from '../types';
 
 const changeConfig = (mode: Mode, config: Readonly<SDKConfig>) => {
