+++ src/lib/visual-editor/msg-router/index.ts
@@ -1,17 +1,3 @@
-import {
-  ChangeConfigEvent,
-  ChangeModeEvent,
-  ChangeVariablesEvent,
-  IncomingEvent,
-  InitEvent,
-} from './incoming-types';
 import { routeIncomingEvent } from './router';
 
-export {
-  routeIncomingEvent,
-  IncomingEvent,
-  InitEvent,
-  ChangeConfigEvent,
-  ChangeModeEvent,
-  ChangeVariablesEvent,
-};
+export { routeIncomingEvent };
