--- src/lib/visual-editor/mutators/types.d.ts
@@ -1,7 +1,0 @@
-type LocalStyles = {
-  [selector: string]: {
-    [property: string]: string;
-  };
-};
-
-export { LocalStyles };
