+++ src/lib/visual-editor/senders/messaging.ts
@@ -1,9 +1,13 @@
 /* eslint-disable functional/no-return-void */
-function postEventToParent(event: object): void {
+
+import { OutboundEvent } from '../types';
+
+function postEventToParent(event: Readonly<OutboundEvent>): void {
   console.log('posting event to parent', event);
   window.parent.postMessage(event, '*');
 }
 
+// TODO: Remove me?
 function postMessageToParent(message: string): void {
   console.log('posting message to parent', message);
   window.parent.postMessage(message, '*');
